+++ src/main/java/com/mycompany/sarisaristore/Product.java
@@ -11,18 +11,18 @@ package com.mycompany.sarisaristore;
 public class Product {
     
     private String name;
-    private int price;
+    private double price;
     private int stock;
     
     
-    public Product(String name, int price, int stock) {
+    public Product(String name, double price, int stock) {
         this.name = name;
         this.price = price;
         this.stock = stock;
     }
     
     public String getName() { return name; }
-    public int getPrice() { return price; }
+    public double getPrice() { return price; }
     public int getStock() { return stock; }
     
     public void updateStock(int stock) {
